requirements.txt
Flask==1.1.1
google-cloud-firestore==1.6.0
google-cloud-bigquery==1.22.0